Reworked CRUD server
Made a bunch of changes here after discussing with Andrej. So
essentially we split the requests up by http method, mapping PUT to
Create, GET to Read, POST to Update, and DELETE to Delete.

GET handles an optional parameter which will return just that record,
or the entire database if no parameter is given.

The other methods are single use.