Added a little express app to test API and get resp from server
Created a little express app in ping/ which uses jquery getJSON function
to call 'GET' on the crud server.

This took some fiddling as there were issues with CORS. This is Cross
Origin Resource Sharing. Generally a browser won't allow loading of
resources from different domains if the request originates from a
script. So because I had crud on localhost:8080 and ping on
localhost:8181, the request wasn't allowed.

The solution is to include an OPTIONS response in the server that
includes headers specifying that the server supports CORS. This OPTIONS
method is called before the actual GET method is called, to check if it
does.