Fixed the server handling JSON from the client
So I figured out that I can just send stringified JSON using XHR. But
this meant changin the server to parse JSON, rather than parsing a
querystring.

This will break the spec as that sends the data as a querystring, so I
need to look at changing the spec to send JSON.

But other than that, working. Also added a couple of event handlers to
log progress on data transfer. At some point would like to add a bigger
db so that progress is actually something worht looking at.